traefik: fix the upgrade problem
Hook resources won't be tracked by Helm. As a result, we'll have to add
a special annotation so that it gets re-created during upgrade. See
more details in https://github.com/helm/helm/issues/5482.

For the traefik case, we also need to re-run the init job for
certificate. This can be achieved by attaching chart release version to
the name of the init job so that it always gets re-run during upgrade.