cert-manager: fix upgrade issue with hooks
Currently, upgrade will fail with "already exists" because the hook
resources are not managed by helm.

To be backwards compatible, we'll have to add the following:
"helm.sh/hook-delete-policy": before-hook-creation

The side effect of this is that the certificates will be re-newed
everything we do upgrade. This is fine because all services including
API server are using root CA for TLS verficiation. Re-newing the
intermediate certs shouldn't affect running services.