chore: bump ingress priority for karma and thanos
The reason is the same as that in
https://github.com/mesosphere/charts/pull/474

When attaching kommander to itself, we observed an issue that the
kommander karma and thanos route (/ops/portal/kommander/monitoring/...,
priority 2) will be overrided by the ops portal overrides (/ops/portal,
priority 3).

Apparently, traefik will first evaluate priority and then do the longest
prefix match on the route path. As a result, we won't be able to see
Kommander karma and thanos UIs after self attaching.

This patch fixed the issue by bumping route priorities to 4.