dka: allow using system default CA
This is for the case when the traefik ingress uses a valid Let's Encrypt
cert that can be trusted by the system default root CA bundles.
Currently, we always display the k8s root CA in the /token page for the
kubectl setup instructions. And k8s root CA cannot validate the Let's
Encrypt cert, leading to authentication errors. See more details in:
https://jira.d2iq.com/browse/D2IQ-65233

This patch introduces an option to use system default CA bundles. This
requires an extra step change in the corresponding repo (coming soon).