bug: fix missing metadata name parameter
This is a hack to work around a deficiency in controller-gen with
sub-components.

With subcomponents, the only option controller-gen gives
is for top-level api types, to mark them with
kubebuilder:validation:EmbeddedResource which creates a
x-kubernetes-embedded-resource marker causing the validator to reject
the resource if it's missing the ApiType and Kind for the subresource.

Since volumeClaimTemplate doesn't actually _require_ those, this causes
the templated resource to fail validation.

This hack is a workaround pending a fix of https://github.com/coreos/prometheus-operator/issues/3093