awsebscsiprovisioner: make upgrade automatic (#610)
* awsebscsiprovisioner: make upgrade automatic

We observed that the upgrade e2e tests on aws always fail with aws csi
driver failing to install v1beta1 volumesnapshot CRDs. This starts to
happen after we bumped the master konvoy to use 1.17 by default.

The issue is caused by this PR
https://github.com/mesosphere/charts/pull/580. On 1.17 clusters, the
chart will try to install v1beta1 CRD. On a 1.16.x cluster, v1alpha1 CRD
exists. The upgrade won't work because only v1beta1 version is listed in
the CRD definition while the storage version is v1alpha1 still.

The above PR states that:
```
atm with this an "upgrade" would fail, and there would be the need to
manual remove the v1alpha1 snapshot CRDs. Which would also purge
snapshots. This must be documented.
Previously, upgrade is manual
```

This patch makes this process automatic.

https://jira.d2iq.com/browse/D2IQ-68987

* chore: stop ebs csi controller during upgrade

This is based on the feedback. v1alpha1 CRDs might be re-installed by
the controller if the pod crashes in the middle of the upgrade. As a
result, we adjusted the hook script to stop the controller before
deleting CRDs.