fix: traefik job name must be unique for upgrades to succeed (#648)
Otherwise, Helm will try to patch the last job and K8s doesn't allow
that for certain fields (e.g. the image field is immutable).

Upon an upgrade of the release comprised of this chart, the previous
job will get removed. Thus, we won't be having old jobs hanging around
on clusters.