fix [staging/velero]: add missing fields to values.yaml
Because the fields were commented in values.yaml, but referenced in the
kubernetes-base-addons values, helm showed a warning:

```
coalesce.go:199: warning: destination for service is a table. Ignoring non-table value <nil>
```

The warning refers to an _internal_, not-publicly-documented helm concept of a
"YAML Table." See https://github.com/helm/helm/issues/4514 for more
information.