Update velero chart to be compatible with helmv3 and helmv2 (#712)
* chore [staging/velero]: Add maintainers, bump version

* chore [staging/velero]: define namespace where it is missing

Helm v2 defaults the namespace to the release namespace, which is
coincidentally correct.

Helm v3 defaults the namespace to the kubeaddons namespace (where
the controller runs), which is incorrect.

* chore [staging/velero]: remove trailing whitespace

* chore [staging/velero]: deploy CRs with 'pre-install' hook

Helm v3 does not install a CR for which there is no matching CRD in the `crds/`
folder. That folder is incompatible with helm v2, so this chart installs CRDs
using a hook.

* chore [staging/velero]: remove CRD templates

Signed-off-by: Daniel Lipovetsky <dlipovetsky@d2iq.com>

* chore [staging/velero]: install CRDs using a Job

Signed-off-by: Daniel Lipovetsky <dlipovetsky@d2iq.com>

* fix [staging/velero]: add missing fields to values.yaml

Because the fields were commented in values.yaml, but referenced in the
kubernetes-base-addons values, helm showed a warning:

```
coalesce.go:199: warning: destination for service is a table. Ignoring non-table value <nil>
```

The warning refers to an _internal_, not-publicly-documented helm concept of a
"YAML Table." See https://github.com/helm/helm/issues/4514 for more
information.

Co-authored-by: Alejandro Escobar <jaescobar.cell@gmail.com>