[prometheus-operator] feat: add scripts to make upgrading easier (#742)
* chore: move all mesosphere files to patch

This moves all the mesosphere-related templates to the patch folder.
This makes it easier to update the operator with a copy and replace as
none of the mesosphere files will get lost.

Signed-off-by: Sam Tran <stran@mesosphere.com>

* feat: add patch files

This adds patch files that copy all the mesosphere files into the
templates/ folder. This way they get deployed as part of the chart.

It makes it easier to keep all the patches in one place and know what
changes are needed/differ from upstream.

Signed-off-by: Sam Tran <stran@mesosphere.com>

* feat: add helpers script for git commands

This adds a helper file so that all the patch files can use the
git_add_and_commit function which adds and commits all the files that
were changed as part of the patch script. This should make it easy to
see the commit history of what changed.

Signed-off-by: Sam Tran <stran@mesosphere.com>

* chore: update prometheus crd patch script

This commit updates the patch script to use a docker image as the
previous code did not run on OSX.

Also, due to the underlying go-parser lib used in the yq docker image,
updating a yaml file automatically reformats the yaml removing empty
newlines and converting multiline strings into one line[1]. This makes
it hard to see the actual needed change for patch7 which is just to
update the crd volumeClaimTemplate.metadata stanza.

Thus, patch7 is changed to apply two commits, one to simply let yq
format the file with no new changes and a second commit with the needed
change. Similar to what we do with chart revision changes.

Also updated the helper script to accept a commit message.

[1] https://github.com/mikefarah/yq/issues/465#issuecomment-643863154

Signed-off-by: Sam Tran <stran@mesosphere.com>

* feat: add update_operator script

This script does a sparse clone of helm/stable/prometheus-operator and
replaces all the files in our fork with latest upstream.

Then it applies all the patches in the /patch folder which should now
make updating our prometheus chart a bit easier.

Signed-off-by: Sam Tran <stran@mesosphere.com>

* chore: add newlines from PR review

Signed-off-by: Sam Tran <stran@mesosphere.com>

* chore: add doc for upgrading + patch comments

Add an UPGRADING doc so users know what to do to upgrade.
Add comments to the patch scripts so users know what each one does.
crd patch script needed explaining.

Signed-off-by: Sam Tran <stran@mesosphere.com>

* chore: add patch for checking CRD exists again

This adds back the check for `monitoring.coreos.com/v1` that was removed
from upstream.

Signed-off-by: Sam Tran <stran@mesosphere.com>

* chore: fixup shellcheck errors

Co-authored-by: Joe Julian <me@joejulian.name>