[prometheus-operator] chore: upgrade to v9.3.1 and patch scripts (#748)"
This actually reverts commit f617bdad35fff95f1dd506c5fb7dcd406ff7a91d.

This adds back the upgrade scripts and upgrades the chart to v9.3.1.

We had to revert this originally because one of the original PRs was
merged without updating the chart version and publishing the chart ended up
overwriting the existing chart on gh-pages, causing it to not deploy
properly on new clusters.