replace deployment when upgrading (#773)
* untar cert-manager chart

* replace deployment when upgrading

The upstream cert-manager chart uses a dynamically generated label in
the selector. The selector is immutable so this fails when upgrading
from tiller to helm v3.

By forking and changing just the deployment name, the deployment will be
recreated using the new selector.

Since the goal is to retire the 0.10 version of cert-manager from our
products in the near future, this should not create an
undue maintenance burden.

* upgrade deprecated apiVersion

* use a version of kubectl that works with the updated CRDs