fix: delete federated resources prior to uninstallation (#832)
All the federated resources are now deleted as part of a pre-delete
hook so that the kubefed controller has the chance to remove the
finalizer from those resources. Without this change, all the federated
resources have always been left behind on clusters when the chart was
deleted because they would have the dangling finalizers on them.

The reason for this is that the controller managing these finalizers
(kubefed controller) is removed before the resources themselves.