create the serviceaccount needed for deletion jobs on delete
When trying to delete kommander, it failed with:

```
Error creating: pods "cleanup-kommander-kubeaddons-ops-portal-username-" is forbidden: error looking up service account kommander/kommander-kubeaddons: serviceaccount "kommander-kubeaddons" not found
```

This was because the install failed (for k8s envioronmental reasons)
which left the helm release in an error state and the service account
not installed.

By adding this hook to the delete phase, this will prevent the cleanup
job from not having its service account.