fix: bump ui to 6.99.4
Updates the UI to use a new socks client library when connecting to
firewalled attached clusters. This fixes and issues where connecting
some firewalled cluster would result in "stuck" status of Loading Data.
This was caused by the socks client not handling timeouts correctly. So
requests would not succeed, fail or timeout.

- Closes D2IQ-78078