fix(cosi-bucket-kit): Correctly parse BucketInfo (#1561)
This fixes a bug that I recently introduced in the parsing of
the BucketInfo value in the COSI secret. This commit refactors
the fetching of the BucketInfo JSON so it does that only once
and then parses the values from that JSON rather than fetching
the secret for every field.